requirements.txt
django==3.1.7
djangorestframework==3.12.2
djangorestframework_simplejwt
django-environ==0.4.5
psycopg2==2.8.6
drf-yasg==1.20.0
requests==2.25.1
gunicorn==20.0.4
clearbit==0.1.7
